Fix VMM inline assembly for 64-bit mode
- Update tlb_flush_all() to use RAX instead of EAX
- Update enable_paging() to accept uint64_t parameter
- Fix CR0 manipulation to use EAX for 32-bit immediate OR
- Update header declaration to match implementation
- Resolves all inline assembly errors in vmm.c

+++ headers/core/memory/vmm/vmm.h
@@ -28,4 +28,4 @@ bool vmm_is_initialized(void);
 void vmm_ensure_initialized(void);
 bool vmm_map_mmio_region(uint32_t physical_addr, uint32_t size);
 
-void enable_paging(uint32_t page_directory_phys_addr);
+void enable_paging(uint64_t page_directory_phys_addr);
